Refactor detect_plate_and_well
The function detect_plate_and_well was changed so that it only
recognizes the following formats:

- plate[plate number/letter]_[well position]
- pl[plate number/letter]_[well position]

The first part is not case-sensitive: "pl" and "plate" can also be
(partially) uppercase.

detect_plate_and_well was simplified to only recognize these formats,
because there are too many different formats for sample names in use
and the regular expression cannot be made to work with every possible
one. Therefore users are now asked to conform to the format described
above if they want to use the 'Upload a plate design' method for
adding sample ID's.