Fix bug in commit 2e64836c
The bug was due to an error in detect_plate_and_well.

Additional issue to be fixed: if detect_plate_and_well fails due
to an unsuitable sample name format, this needs to be communicated
to the user. The current situation is that the "Add sample ID's"
button is unclickable, but it is unclear why.