Implement new spectra curation methods
In a meeting with David and Tamas we decided on 3 options
for spectra curation methods: based on negative controls,
based on percentiles and skip spectra curation.
In this commit the first two options were implemented.
For skip spectra curation the dataframe that is passed along to
analyte curation still needs to be made suitable for the next
step.