Change .fa to .fas
In FontAwesome 6 the class of the icons is
changed from "fa" to "fas". For unclear reason
FontAwesome 6 is used in the version of the
dashboard on server, even though version 5 is
used in the source code. Styling is no longer
applied to the icons because the class name has
changed.